refactor PPQConnect interfaces
