+++ src/frontend/src/components/public/PPQConnect/interfaces.ts
@@ -1,0 +1,5 @@
+import { IPPQResult } from '../../../types/interfaces/ppq-result.interface';
+
+export interface IComponentProps {
+  result: IPPQResult;
+}
